pyproject.toml
[project]
name = "ai-sales-trainer-poc"
version = "0.1.0"
description = "AI Sales Trainer PoC - RAG, Voice Agent, and Evaluation"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn[standard]>=0.40.0",
    "openai>=2.15.0",
    "qdrant-client>=1.16.0",
    "pypdf>=6.6.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.1.0",
    "google-genai>=1.56.0",
    "boto3>=1.42.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
